% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{anxiety_fun}
\alias{anxiety_fun}
\title{Range anxiety}
\usage{
anxiety_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real anxiety parameter for bevs. 1 is rational range anxiety, > 1 is irrational
}
\description{
Range anxiety
}
\examples{
anxiety_fun(scenario_B,2023)
}
