% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{battery_cost_function}
\alias{battery_cost_function}
\title{Battery cost function}
\usage{
battery_cost_function(sD, yeartime, nlag = 4.5)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time 2015-2050}

\item{nlag}{lag in years between bnef battery price and production price}
}
\value{
real (euros)
}
\description{
Learning curve model based on historical BNEF survey prices with a production lag and a learning curve
 i.e. early 2019 production models reflect mid 2016 BNEF battery prices
}
