% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{bev_fuelcost}
\alias{bev_fuelcost}
\title{bev_fuelcost}
\usage{
bev_fuelcost(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real euros per km
}
\description{
energy cost of bev in euros/km
}
