% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bev_historical_uptake.R
\docType{data}
\name{bev_historical_uptake}
\alias{bev_historical_uptake}
\title{BEV Penetration}
\format{
A data frame with 38 rows and 2 variables:
\describe{
\item{date}{year}
\item{penetration}{BEV penetration at year end}
}
}
\source{
<cso>
}
\usage{
bev_historical_uptake
}
\description{
Historical BEV penetration at year end inferred from CSO i.e. fraction of on-the-road private BEVs
}
\keyword{datasets}
