% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{bev_premium_fun}
\alias{bev_premium_fun}
\title{BEV price premium (euros)}
\usage{
bev_premium_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
BEV price premium (euros)
}
