% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{bev_rrp_full}
\alias{bev_rrp_full}
\title{rrp_full}
\usage{
bev_rrp_full(rrp_incentive, params)
}
\arguments{
\item{rrp_incentive}{list price (euros) of bev including incentive (grant + vrt rebate)}

\item{params}{hashed parameters}
}
\value{
price (euros)
}
\description{
full list price of a bev before incentives
}
