% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget_vals.R
\docType{data}
\name{budget_vals}
\alias{budget_vals}
\title{qev32 integer to budget value}
\format{
A vector:
\describe{
\item{}{}
}
}
\source{
2018 survey range midpoints
}
\usage{
budget_vals
}
\description{
qev32 integer to budget value
}
\keyword{datasets}
