% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{car_vat_fun}
\alias{car_vat_fun}
\title{VAT on cars}
\usage{
car_vat_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
real vat rate
}
\description{
VAT on cars
}
