% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{carbon_tax_fun}
\alias{carbon_tax_fun}
\title{Carbon tax}
\usage{
carbon_tax_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
real
}
\description{
Carbon tax
}
