% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{conversion_fun}
\alias{conversion_fun}
\title{conversion_fun}
\usage{
conversion_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
value of interpolated conversion efficiency in kWh/km at yeartime
}
\description{
conversion_fun
}
\examples{
conversion_fun(scenario_B,2026.5)

}
