% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{depreciation_spread_fun}
\alias{depreciation_spread_fun}
\title{ev_depreciation_fun}
\usage{
depreciation_spread_fun(type, sD, yeartime)
}
\arguments{
\item{type}{fuel/powertrain type}

\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
percentage spread as fraction
}
\description{
spread between zevs & hev diesel/petrol
}
