% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{diesel_excise_duty_fun}
\alias{diesel_excise_duty_fun}
\title{Diesel Excise Duty}
\usage{
diesel_excise_duty_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
excise euros/litre at yeartime
}
\description{
diesel excise scenario based on
}
\examples{
diesel_excise_duty_fun(scenario_1,2030)
}
