% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{diesel_price}
\alias{diesel_price}
\title{diesel_price}
\usage{
diesel_price(
  crudeoilprice = 50,
  crackspread = 10,
  margin = 0.12,
  carbon_tax = 33.5,
  excise_duty = 0.54,
  NORA = 0.02,
  bio = 0.014,
  VAT = 0.23
)
}
\arguments{
\item{crudeoilprice}{euros/barrel}

\item{crackspread}{euros}

\item{margin}{euros}

\item{carbon_tax}{euro/tCO2}

\item{excise_duty}{euros}

\item{NORA}{levy euros}

\item{bio}{levy euros}

\item{VAT}{percent}
}
\value{
euros per litre
}
\description{
Retail fuel price formula. Does not distinguish between petrol and diesel.
}
