% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phev.R
\name{e_ice}
\alias{e_ice}
\title{e_ice}
\usage{
e_ice(wltp, AER)
}
\arguments{
\item{wltp}{WLTP quoted emissions for phev}

\item{AER}{all electric range}
}
\value{
gCO2/km
}
\description{
emissions from a charged depleted phev based in European utility factor
}
