% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phev.R
\name{e_phev}
\alias{e_phev}
\title{e_phev}
\usage{
e_phev(wltp, AER, mileage, xi = 1)
}
\arguments{
\item{wltp}{WLTP quoted emissions for phev}

\item{AER}{all electric range}

\item{mileage}{km from mileage_vals}

\item{xi}{charging behaviour 0.25,0.5,0.75,1}
}
\value{
gCO2/km
}
\description{
real world phev emissions model based on European utility factor daily charging fraction xi
}
