% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{electricity_price_fun}
\alias{electricity_price_fun}
\title{Electricity price euros/kWh}
\usage{
electricity_price_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
Electricity price euros/kWh
}
\examples{
electricity_price_fun(scenario_B,2029)

}
