% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_model.R
\name{emissions_model}
\alias{emissions_model}
\title{ICEV specific emissions}
\usage{
emissions_model(mileage, e_upper = 180, e_lower = 80)
}
\arguments{
\item{mileage}{real}

\item{e_upper}{real}

\item{e_lower}{real}
}
\value{
real
}
\description{
Statistical Model for specific ICEV emissions depending on mileage. In 2015 fleet mean NEDC emissions sugested by motor tax data is 150gCO2/km.
}
