% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{emissions_model1}
\alias{emissions_model1}
\title{emissions_model1}
\usage{
emissions_model1(
  mileage,
  mean_emissions = 190,
  sd_emissions = 30,
  e_cut = 80,
  zeta = 0.75
)
}
\arguments{
\item{mileage}{km}

\item{mean_emissions}{gco2/km - corresponds to wltm emissions  e.g. 190gCO2/km in 2015}

\item{sd_emissions}{gco2/km -}

\item{e_cut}{lower bound on ICEV emissions, roughly set to Toyota Prius wltp emissions}

\item{zeta}{determines strength of correlation between mileage and emissions}
}
\value{
a randomly generated specific emissions in gCO2/km
}
\description{
gamma distribution of emissions above a cutoff corresponding to input mean emissions and standard deviation parameters. Dependence on mileage is modelled by a tanh function.
}
\examples{

}
