% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{excise_duty_fun}
\alias{excise_duty_fun}
\title{Excise Duty}
\usage{
excise_duty_fun(type, sD, yeartime)
}
\arguments{
\item{type}{diesel or gasoline}

\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
excise euros/litre at yeartime
}
\description{
historical and future motor fuel excise (in euros)
}
\examples{
excise_duty_fun("gasoline",scenario_B,2030)
}
