% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{fleet_factor_fun}
\alias{fleet_factor_fun}
\title{fleet_factor_fun}
\usage{
fleet_factor_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
emissions relative to 2021 value
}
\description{
ICEV fleet emissions normalises to 2021 fleet WLTP values.
}
