% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{fuelcost_fun}
\alias{fuelcost_fun}
\title{fuelcost_fun}
\usage{
fuelcost_fun(type, wltp = 0, kWh, AER, params, xi = 1)
}
\arguments{
\item{type}{diesel,petrol,hev,phev,bev}

\item{wltp}{emissions g/km}

\item{kWh}{battery capacity}

\item{AER}{all electric range km}

\item{params}{current value of scenario params from scenario_params()}

\item{xi}{charging behaviour}
}
\value{
real euros/km
}
\description{
ICEV fuel cost in euros/km
}
