% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{gasoline_excise_duty_fun}
\alias{gasoline_excise_duty_fun}
\title{Diesel Excise Duty}
\usage{
gasoline_excise_duty_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
excise euros/litre at yeartime
}
\description{
Gasoline excise scenario based on
}
\examples{
gasoline_excise_duty_fun(scenario_1,2025)
}
