% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_current_tech}
\alias{get_current_tech}
\title{get_current_tech}
\usage{
get_current_tech(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
current fleet data frame
}
\description{
new fleet parameters at current (pre or post 2021) times derived from adjusted 2021 fleet prices and tech. 
Adjustments made include falling battery pack prices for EVs, efficiency (AER adjustments), ICEVs specific emissions reductions etc
only models with model_start < yeartime are included
}
