% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_newtech}
\alias{get_newtech}
\title{get_newtech}
\usage{
get_newtech(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
new car fleet at yeartime
}
\description{
New technology dataframe available at yeartime based on 2021 and 2022 new car fleets.
The 2021 fleet is used to project backwards and the 2022 fleet is used to project forwards in time
}
