% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_newtech_2021}
\alias{get_newtech_2021}
\title{get_newtech_2021}
\usage{
get_newtech_2021(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
current fleet data frame
}
\description{
past (pre 2022) new fleet parameters derived from adjusted 2021 fleet prices and tech.
Adjustments made include falling battery pack prices for EVs, efficiency (AER adjustments), ICEVs specific emissions reductions etc
models with model_start > yeartime are excluded. separate deflation rates of prices for new tech (hev, phev and bevs) are allowed
}
