% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_newtech_2022}
\alias{get_newtech_2022}
\title{get_newtech_2022}
\usage{
get_newtech_2022(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
a new car fleet data frame
}
\description{
future (post 2021) new fleet parameters derived by adjusting adjusting 2022 fleet prices and tech.
Adjustments include falling battery pack prices for EVs, efficiency (AER adjustment), ICEV specific emissions reductions (described by fleet_factor_fun).
}
