% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_used_tech}
\alias{get_used_tech}
\title{get_used_tech}
\usage{
get_used_tech(sD, yeartime, vehicle_age = 4)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal purchase time of vehicle}

\item{vehicle_age}{age of vehicle (decimal time)}
}
\value{
current fleet data frame
}
\description{
available used fleet parameters at future or past times based on adjusted new fleet prices and tech.
adjustments made include falling battery pack prices for EVs, efficiency (AER adjustments), ICEVs specific emissions reductions etc
only models with model_start < yeartime are included
}
