% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getActivity}
\alias{getActivity}
\title{getActivity}
\usage{
getActivity(abm)
}
\arguments{
\item{abm}{output of runABM}
}
\value{
dataframe type, year, activity
}
\description{
annual activity (km) by powertrain type
}
