% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getAdoption}
\alias{getAdoption}
\title{getAdoption}
\usage{
getAdoption(abm, addvars = NULL)
}
\arguments{
\item{abm}{output ofrunABM}

\item{addvars}{additional grouping variable (default is type only)}
}
\value{
uptake dataframe
}
\description{
getAdoption
}
