% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getEmissions}
\alias{getEmissions}
\title{getEmissions}
\usage{
getEmissions(
  abm,
  addvars = NULL,
  xi = 1,
  average_over_runs = F,
  emissions_start_date = NULL
)
}
\arguments{
\item{abm}{output of runABM}

\item{addvars}{vector of additional variables}

\item{xi}{phev behavioural parameter xi=1 is optimal charging behaviour}

\item{average_over_runs}{if FALSE return individual run emissions, otherwise average over runs}

\item{emissions_start_date}{start date for cumulative emissions calculation}
}
\value{

}
\description{
Monthly and cumulative emissions for a set of model runs
}
