% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getEVAdoption1}
\alias{getEVAdoption1}
\title{BEV Adoption Rates}
\usage{
getEVAdoption1(abm)
}
\arguments{
\item{abm}{agent dataframe by simulation run and date}
}
\value{
dataframe columns simulation (run ID) - date - predicted uptake fraction
}
\description{
Converts ABM output to EV adoption rates by simulation run
}
