% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getFleetAge}
\alias{getFleetAge}
\title{get fleet age}
\usage{
getFleetAge(abm, addvars = NULL)
}
\arguments{
\item{abm}{output of runABM}

\item{addvars}{additional variable (type,segment, qev31 etc)}
}
\value{

}
\description{
get fleet age
}
