% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getIncentiveCost}
\alias{getIncentiveCost}
\title{getIncentiveCost}
\usage{
getIncentiveCost(abm)
}
\arguments{
\item{abm}{simulation runs (output of runABM)}
}
\value{
incentive cost (difference between after tax price and incentivised price)
}
\description{
getIncentiveCost
}
