% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getMileages}
\alias{getMileages}
\title{getMileages}
\usage{
getMileages(abm)
}
\arguments{
\item{abm}{output of runABM}
}
\value{
datafame of date, type and mean annual mileage
}
\description{
mileage by fuel type and other variables if requires
}
