% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getZEVSales}
\alias{getZEVSales}
\title{getZEVSales}
\usage{
getZEVSales(abm)
}
\arguments{
\item{abm}{output of runABM}
}
\value{
sales dataframe (year, sales fraction)
}
\description{
annual ZEV sales market share by
}
