% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{grant_fun}
\alias{grant_fun}
\title{Grant (linear phaseout)}
\usage{
grant_fun(type, sD, yeartime)
}
\arguments{
\item{type}{car type bev, phev, hev, diesel, petrol}

\item{sD}{scenarion dataframe}

\item{yeartime}{decimal time}
}
\value{
grant level at yeartime. 1 is current grant level, 0 is no grant
}
\description{
Grant (linear phaseout)
}
