% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{grant_fun1}
\alias{grant_fun1}
\title{grant_fun1}
\usage{
grant_fun1(type, sD, yeartime)
}
\arguments{
\item{type}{phev or bev}

\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
grant phaseout path usually in 0,1
}
\description{
grant phaseout scenario for phevs and bevs
}
