% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{graph_adoption}
\alias{graph_adoption}
\title{Graph Model Uptake}
\usage{
graph_adoption(
  abm,
  yscale = "sqrt",
  gtitle = "EV number",
  galpha = 0.4,
  fleet_size = 2200
)
}
\arguments{
\item{abm}{output from runABM}

\item{yscale}{y-scale "sqrt" or "linear"}

\item{gtitle}{graph title}

\item{galpha}{alpha transparency for graph}

\item{fleet_size}{assumed constant motor fleet number in thousands}
}
\value{
ggplot object
}
\description{
Make a graph including historical data from bev_historical_uptake
}
