% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily-data.R
\docType{data}
\name{homophily}
\alias{homophily}
\title{Agent homophily}
\format{
A data frame with 11 rows and 2 variables:
\describe{
\item{variable}{demographic variable}
\item{weights}{homophily weight}
...
}
}
\source{
experiments and judgement
}
\usage{
homophily
}
\description{
Demographic variables used build homophilous network
}
\keyword{datasets}
