% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_model.R
\name{icev_fleet_fun}
\alias{icev_fleet_fun}
\title{icevfleet_fun}
\usage{
icev_fleet_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
ICEV fleet upper emissions parameter
}
