% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_model.R
\name{icev_fleet_lowercut_fun}
\alias{icev_fleet_lowercut_fun}
\title{icev_fleet_lowercut_fun}
\usage{
icev_fleet_lowercut_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time in years}
}
\value{
value of e_cut used in emissions_model1
}
\description{
lower bound on ICEV emission (i.e. non-plug-in hyrbids)
}
