% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{icevfleet_fun}
\alias{icevfleet_fun}
\title{ICEV fleet emissions upper parameter}
\usage{
icevfleet_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
ICEV fleet emissions upper parameter
}
