% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{incentives_fun}
\alias{incentives_fun}
\title{incentives_fun}
\usage{
incentives_fun(type, p_1, params, wltp, new_used)
}
\arguments{
\item{type}{fuel/powertrain}

\item{p_1}{unsubsidised price (including VAT and VRT)}

\item{params}{hashed parameter values}

\item{wltp}{gCO2/km}

\item{new_used}{new or used vehicle, values "new_car" or "used_car"}
}
\value{
incentive in euros
}
\description{
total subsidy (SEAI grant + VRT rebate)
}
