% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_fleet.R
\docType{data}
\name{initial_fleet}
\alias{initial_fleet}
\title{initial_fleet}
\format{
A data frame with 225 rows and 5 variables:
\describe{
  \item{rank}{ranl}
  \item{make}{manufacturer}
  \item{model}{model}
  \item{number}{number purchased}
  \item{segment}{segment}
}
}
\source{
\url{www.cso.ie}
}
\usage{
initial_fleet
}
\description{
new vehicle models purchased 2014-2019
}
\keyword{datasets}
