% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{initialise_segments}
\alias{initialise_segments}
\title{initialise_segments}
\usage{
initialise_segments(sD)
}
\arguments{
\item{sD}{scenario (needed for historical fleet parameters)}
}
\value{
initial agents dataframe
}
\description{
extends agents_init with vehicle model & segment choice based on naive bayes classifier
A car model is the assigned by on segment choice, observed Irish make preferences 2014-2019 & 2021, and 2021 new fleet dataframe
}
