% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mileage_vals.R
\docType{data}
\name{mileage_vals}
\alias{mileage_vals}
\title{qev34 integer to mileage value}
\format{
A vector:
\describe{
\item{}{}
}
}
\source{
2018 survey range midpoints
}
\usage{
mileage_vals
}
\description{
qev34 integer to mileage value
}
\keyword{datasets}
