% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{motor_tax}
\alias{motor_tax}
\title{motor_tax}
\usage{
motor_tax(emissions, flag)
}
\arguments{
\item{emissions}{WLTP gCO2/km}

\item{flag}{old or new vrt}
}
\value{
annual amount in euros
}
\description{
WLTP emissions based motor tax post Jan 2021
}
