% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{motor_tax_old}
\alias{motor_tax_old}
\title{Annual Motor Tax}
\usage{
motor_tax_old(emissions)
}
\arguments{
\item{emissions}{WLTP emissions gCO2/km}
}
\value{
euros
}
\description{
This function looks up tax band in motortaxbands depending on WLTP emissions
}
