% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motortaxbands.R
\docType{data}
\name{motortaxbands}
\alias{motortaxbands}
\title{VRT and Motor Tax}
\format{
A data frame with 49 rows and 6 variables:
\describe{
  \item{name}{variable type}
  \item{lower}{variable description}
  \item{upper}{name of variable}
  \item{motor}{scenario value}
  \item{rate_vrt}{units}
  \item{min_vrt}{source etc}
}
}
\source{
\url{www.motortax.ie}
}
\usage{
motortaxbands
}
\description{
Emissions based bands up to  Jan 1 2021
}
\keyword{datasets}
