% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_motortax_bands.R
\docType{data}
\name{new_motortax_bands}
\alias{new_motortax_bands}
\title{VRT WLTP bands and rates from Jan 1 2020}
\format{
A data frame with 21 rows and 4 variables:
\describe{
  \item{band}{number}
  \item{lower}{lower limit}
  \item{upper}{upper limit}
  \item{motor}{annual payment in euros}

}
}
\source{
\url{revenue.ie}
}
\usage{
new_motortax_bands
}
\description{
Imported NEDC rated cars will be charged VRT based on transformation to WLTP
}
\keyword{datasets}
