% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{oil_price_fun}
\alias{oil_price_fun}
\title{Oil price scenario}
\usage{
oil_price_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
Oil price scenario
}
