% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opec_oilprices.R
\docType{data}
\name{opec_oilprices}
\alias{opec_oilprices}
\title{Mean Annual OPEC Crude Oil Price}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{year}{year}
\item{price}{$/barrel}
\item{EUR}{EURUSD annual mean}
\item{price_euro}{euro/barrel}
}
}
\source{
<www.statistica.com>
}
\usage{
opec_oilprices
}
\description{
Mean Annual OPEC Crude Oil Price
}
\keyword{datasets}
