% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{range_factor_fun}
\alias{range_factor_fun}
\title{range_factor_fun}
\usage{
range_factor_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
unitless
}
\description{
BEV range function factor models efficiency improvements over time (i.e. km/kWh)
}
