% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{runABM}
\alias{runABM}
\title{runABM}
\usage{
runABM(
  sD,
  Nrun = 1,
  simulation_end = end_year,
  resample_society = F,
  n_unused_cores = 2
)
}
\arguments{
\item{sD}{scenario set-up dataframe, typically read with readxlxs(...,sheet=scenario)}

\item{Nrun}{integer, number runs}

\item{simulation_end}{the final year of simulation of early termination is required}

\item{resample_society}{if TRUE resample society with replacement to capture additional variability}

\item{n_unused_cores}{number of cores left unused in parallel/foreach. Recommended values 2 or 1.}
}
\value{
a three component list - simulation output, scenario setup, meta-parameters
}
\description{
Runs the simulation on artificial society of 924 agents. Each run is performed on an independently generated social network with random initial ICEV emissions
}
