% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_1.R
\docType{data}
\name{scenario_1}
\alias{scenario_1}
\title{scenario_1}
\format{
A data frame with 225 rows and 5 variables:
\describe{
\item{category}{variable category}
\item{variable}{variable description}
\item{parameter}{parameter name}
\item{value}{value assumed}
\item{unit}{value units}
\item{comment}{comment}

}
}
\source{
<various>
}
\usage{
scenario_1
}
\description{
moderate policy, favourable technology scenario.
}
\keyword{datasets}
