% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_2.R
\docType{data}
\name{scenario_2}
\alias{scenario_2}
\title{scenario_2}
\format{
A data frame with 225 rows and 5 variables:
\describe{
  \item{category}{variable category}
  \item{variable}{variable description}
  \item{parameter}{parameter name}
  \item{value}{value assumed}
  \item{unit}{value units}
  \item{comment}{comment}
  
}
}
\source{
\url{various}
}
\usage{
scenario_2
}
\description{
moderate policy, favourable technology scenario, early phev grant phaseout
}
\keyword{datasets}
