% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_3.R
\docType{data}
\name{scenario_3}
\alias{scenario_3}
\title{scenario_3}
\format{
A data frame with 225 rows and 5 variables:
\describe{
  \item{category}{variable category}
  \item{variable}{variable description}
  \item{parameter}{parameter name}
  \item{value}{value assumed}
  \item{unit}{value units}
  \item{comment}{comment}
  
}
}
\source{
\url{various}
}
\usage{
scenario_3
}
\description{
pump price increases to euro 2.40, favourable technology scenario, bev grant maintained until 2025
}
\keyword{datasets}
