% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_B.R
\docType{data}
\name{scenario_B}
\alias{scenario_B}
\title{scenario_B}
\format{
A data frame with 225 rows and 5 variables:
\describe{
\item{category}{variable category}
\item{variable}{variable description}
\item{parameter}{parameter name}
\item{value}{value assumed}
\item{unit}{value units}
\item{comment}{comment}

}
}
\source{
<various>
}
\usage{
scenario_B
}
\description{
a scenario with €2.30 pump prices in 2030.
}
\keyword{datasets}
