% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{scenario_fulldata}
\alias{scenario_fulldata}
\title{scenario_fulldata}
\usage{
scenario_fulldata(sD)
}
\arguments{
\item{sD}{scenario dataframe}
}
\value{
parameter set
}
\description{
scenario developed as a function of date for plotting
}
