% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{scenario_params}
\alias{scenario_params}
\title{scenario_params}
\usage{
scenario_params(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
environment
}
\description{
hashed for fast lookup
}
