% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{scenario_params_df}
\alias{scenario_params_df}
\title{scenario_params_df}
\usage{
scenario_params_df(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
object of class "environment" for fast lookup e.g. params$battery_cost
}
\description{
techno-economic parameter dataframe at yeartime
parameters derived from a scenario at a decimal time between 2015 and 2050. Used to create hash object for fastlookup of current parameter values.
}
