% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{tech_cost_fun}
\alias{tech_cost_fun}
\title{tech_cost_fun}
\usage{
tech_cost_fun(type, tech_cost, kWh = NA, params)
}
\arguments{
\item{type}{powertrain/fuel type}

\item{tech_cost}{2021 technology cost}

\item{kWh}{battery}

\item{params}{scenario params e.g. scenario_1}
}
\value{
cost in euros
}
\description{
technology costs based on 2021 pre-tax and incentive list prices. 
At present the cost model is based on battery pack price changes only.
}
