% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{tech_deflate}
\alias{tech_deflate}
\title{tech_deflate}
\usage{
tech_deflate(rate, yeartime)
}
\arguments{
\item{rate}{price deflation rate}

\item{yeartime}{decimal time}
}
\value{
real value (1 in 2021)
}
\description{
helper function describing for ev price factors pre-2021
}
