% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phev.R
\name{uf}
\alias{uf}
\title{uf}
\usage{
uf(AER)
}
\arguments{
\item{AER}{all electric range (e.g. WLTP)}
}
\value{
utilisation factor 0-1
}
\description{
utility factor from phev all electric range
https://circabc.europa.eu/sd/a/92324676-bd8c-4075-8301-6caf12283beb/Technical%20Report_UF_final.pdf
}
