% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_interp_data_xi_1.0.R
\docType{data}
\name{uf_interp_data_xi_1.0}
\alias{uf_interp_data_xi_1.0}
\title{phev utility factors (uf) for use in uf_phev_interp}
\format{
A matrix with 200 rows (AERs) and 8 columns (mileages)
}
\source{
internal
}
\usage{
uf_interp_data_xi_1.0
}
\description{
Computed at 200 AERs and 8 mileages corresponding to xi=1.0
}
\keyword{datasets}
