% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_model_wide.R
\docType{data}
\name{uf_model_wide}
\alias{uf_model_wide}
\title{Parameters of mileage and xi dependent PHEV utility factor model}
\format{

}
\source{
{generated by uf_simulation.R}
}
\usage{
uf_model_wide
}
\description{
Parameters of mileage and xi dependent PHEV utility factor model
}
\keyword{datasets}
