% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phev.R
\name{uf_phev}
\alias{uf_phev}
\title{uf_phev}
\usage{
uf_phev(mileage1, xi1, AER)
}
\arguments{
\item{mileage1}{mileage category km}

\item{xi1}{charging behaviour parameter}

\item{AER}{PHEV range}
}
\value{
utility factor in 0,1
}
\description{
uf_phev
}
