% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phev.R
\name{uf_phev_interp}
\alias{uf_phev_interp}
\title{uf_phev_interp}
\usage{
uf_phev_interp(mileage1, xi1, AER)
}
\arguments{
\item{mileage1}{kms, any mileage <=60000}

\item{xi1}{charging behaviour allowed values 0.5,1,1.5,2}

\item{AER}{electric range km}
}
\value{
real
}
\description{
fast, interpolated phev utilisation factor
}
