% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{update_agents4}
\alias{update_agents4}
\title{update_agents4}
\usage{
update_agents4(sD, yeartime, agents_in, social_network)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}

\item{agents_in}{input agent dataframe}

\item{social_network}{artifical social network}
}
\value{
new agent dataframe
}
\description{
Micro-simulation Updater
}
\details{
The workhorse ABM function.Within a scenario, does a single month update of the agent characteristics. A random sample of agents evaluates their economic and social
utilities. If these exceed their individual threshold and EV is adopted.
}
