% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_empirical.R
\docType{data}
\name{util_empirical}
\alias{util_empirical}
\title{Empirical utilites from Survey Model}
\format{
A data frame with 25 rows and 4 variables:
\describe{
\item{name}{survey question codes: q17a_2 = environmental concern, qev29 = number of EV owners known}
\item{answercode}{feature value}
\item{answer}{answer}
\item{shap}{SHAP contribution to utility prediction}
}
}
\source{
<Survey Model>
}
\usage{
util_empirical
}
\description{
SHAP values for environmental, social, budget and mileage
}
\keyword{datasets}
