% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vkt_factor_fun}
\alias{vkt_factor_fun}
\title{vkt_factor}
\usage{
vkt_factor_fun(price_elasticity, price, demand_reduction)
}
\arguments{
\item{price_elasticity}{travel demand elasticity}

\item{price}{prevailing price future fuel price}

\item{demand_reduction}{prevailing demand_reduction demand reduction due to non-price measures - active travel, congestion charging etc e.g. 0.95}
}
\value{
vkt factor 0-1
}
\description{
a simple mileage reduction factor based on fuel price and demand reduction measures
}
