% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{vkt_reduction_fun}
\alias{vkt_reduction_fun}
\title{VKT reduction}
\usage{
vkt_reduction_fun(sD, mileage, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{mileage}{km}

\item{yeartime}{decimal time}
}
\value{
dimensionless fraction
}
\description{
Reduction factor of vehicle kilometers travelled as a function of mileage
}
