% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt}
\alias{vrt}
\title{vrt}
\usage{
vrt(price, emissions, flag)
}
\arguments{
\item{price}{omsp euros}

\item{emissions}{WLTP emissions gCO2/km}

\item{flag}{old or new vrt}
}
\value{
amount in euros
}
\description{
VRT amount paid. At the moment this assumes the old minimum VRT amounts
}
