% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_bands_2021.R
\docType{data}
\name{vrt_bands_2021}
\alias{vrt_bands_2021}
\title{VRT WLTP bands and rates from Jan 1 2020}
\format{
A data frame with 21 rows and 4 variables:
\describe{
\item{band}{number}
\item{lower}{lower limit}
\item{upper}{upper limit}
\item{rate}{VRT rate}

}
}
\source{
<revenue.ie>
}
\usage{
vrt_bands_2021
}
\description{
Imported NEDC rated cars will be charged VRT based on transformation to WLTP
}
\keyword{datasets}
