% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt_rate}
\alias{vrt_rate}
\title{vrt_rate}
\usage{
vrt_rate(emissions, flag)
}
\arguments{
\item{emissions}{WLTP gCO2/km}

\item{flag}{old or new vrt}
}
\value{
rate
}
\description{
VRT rates based on old/new flag
}
