% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt_rebate}
\alias{vrt_rebate}
\title{vrt_rebate}
\usage{
vrt_rebate(type, omsp, params, flag)
}
\arguments{
\item{type}{vehicle type bev,phev, hev, diesel, petrol}

\item{omsp}{approximated by budget + dP_0}

\item{params}{hashed parameter object}

\item{flag}{old or new basis}
}
\value{
amount in euros
}
\description{
post 2020 VRT rebate amount for BEVs based on OMSP
}
