% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt_rebate1}
\alias{vrt_rebate1}
\title{vrt_rebate1}
\usage{
vrt_rebate1(type, omsp, params, flag, WLTP, new_used = "new_car")
}
\arguments{
\item{type}{fuel/powertain}

\item{omsp}{open maket sale price in euros}

\item{params}{current scenario parameters generated from scenario_params}

\item{flag}{new or old tax basis (pre or post Jan 1 2021)}

\item{WLTP}{gCO2/km}

\item{new_used}{new or used vehicle "new_car" or "used_car"}
}
\value{

}
\description{
vrt relief function
}
