% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt_rebate_lower}
\alias{vrt_rebate_lower}
\title{vrt_rebate_lower}
\usage{
vrt_rebate_lower(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
euros
}
\description{
The upper rebate window, set to 40,000 euros in 2021.
}
