% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{vrt_rebate_upper}
\alias{vrt_rebate_upper}
\title{vrt_rebate_upper}
\usage{
vrt_rebate_upper(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
euros
}
\description{
The lower rebate window, set to 10,000 euros in 2021. VRT rebate is 50% between vrt_rebate_lower and vrt_rebate_lower + vrt_rebate_upper
}
