% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zevgrantbands.R
\docType{data}
\name{zevgrantbands}
\alias{zevgrantbands}
\title{BEV grants (2020)}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{min}{lower price range, euros}
\item{max}{upper price range, euros}
\item{grant}{grant in euros}
}
}
\source{
\url{http://www.seai.ie/}
}
\usage{
zevgrantbands
}
\description{
Grants from SEAI
}
\keyword{datasets}
