% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnefprices.R
\docType{data}
\name{bnefprices}
\alias{bnefprices}
\title{BNEF battery pack prices}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{year}{survey year}
  \item{pack_$kWh}{price per kWh in USD}
  \item{EUR}{mean annual exchange rate}
  \item{pack_ekWh}{price in EUR}
}
}
\source{
\url{https://cleantechnica.com/}
}
\usage{
bnefprices
}
\description{
Emissions based bands up to  Jan 1 2021
}
\keyword{datasets}
