% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{crackspread_fun}
\alias{crackspread_fun}
\title{crackspread_fun}
\usage{
crackspread_fun(type, sD, yeartime)
}
\arguments{
\item{type}{diesel or gasoline}

\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
euros/barrel
}
\description{
refining margin for diesel or gasoline
}
