% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{fleet_factor_fun}
\alias{fleet_factor_fun}
\title{fleet_factor_fun}
\usage{
fleet_factor_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
emissions relative to 2021 value
}
\description{
ICEV fleet emissions technology improvement based on \% reduction in WLTP values.
This is applied to each ICEV model to project it forward and backwards in time.
Backward projection from 2021, forward from 2022 i.e. fleet_factor = 1 for 2021 and 2022
}
