% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{oil_price_fun}
\alias{oil_price_fun}
\title{Oil price scenario}
\usage{
oil_price_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{real}
}
\value{
real
}
\description{
continuous time oil price (euro/bbl) based on monthly opec basket price (ORB) and monthly eurusd settlements from ecb
this depends on assumed oil prices at of 2030 and 2050 from input scenario
}
\examples{
oil_price_fun(scenario_B,2023)

}
