% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_model_wide.R
\docType{data}
\name{uf_model_wide}
\alias{uf_model_wide}
\title{Parameters of mileage and xi dependent PHEV utility factor model}
\format{
A data frame with 56 rows and 13 variables:
\describe{
\item{xi}{charging behaviour parameter with values 1,0.75,0.5,0.25. 1 means 1 full charge per day, 0.5 means random full charge 50\% of days etc}
\item{mileage}{mileage category in km}
\item{c_0}{exponential model parameter}
\item{c_1}{exponential model parameter}
\item{c_2}{exponential model parameter}
\item{c_3}{exponential model parameter}
\item{c_4}{exponential model parameter}
\item{c_5}{exponential model parameter}
\item{c_6}{exponential model parameter}
\item{c_7}{exponential model parameter}
\item{c_8}{exponential model parameter}
\item{c_9}{exponential model parameter}
\item{c_10}{exponential model parameter}
}
}
\source{
{generated by uf_simulation.R}
}
\usage{
uf_model_wide
}
\description{
Parameters of mileage and xi dependent PHEV utility factor model
}
\keyword{datasets}
